Clear color can now be adjusted in camera properties

+++ Phoenix-Editor/src/Panels/ContentBrowserPanel.cpp
@@ -181,7 +181,7 @@ namespace phx {
 		ImGui::Begin("Content Browser");
 
 		ImGui::PushStyleColor(ImGuiCol_Button, ImVec4(0.0f, 0.0f, 0.0f, 0.0f));
-		int size = ImGui::GetFontSize();
+		float size = ImGui::GetFontSize();
 		
 		UI::PushHiddenButton();
 		if (ImGui::Button("<", ImVec2(size, size)))
